task7/app/requirements.txt
psycopg2-binary==2.9.9
prompt-toolkit==3.0.48
fastapi
uvicorn
jinja2
python-multipart